protocol: reduce the yield time in QMK thread to 1 us
The traditional QMK main is not running in a RTOS thread, thus cannot yield
itself, and other newly added thread can get get resources to complete their
tasks. The example we have encountered is that BLE cannot complete the BT SMP
process due to insufficient CPU time slot.

Before, we force QMK main thread to yield with sleeping 1 ms in
`protocol_pre_task` and `protocol_post_task`. After profiling the thread usage,
sleeping 1 ms would cause the system in the idle state for ~93% of the CPU time.
Reducing the sleep time to 1 us results the idle state for ~33% of the CPU time,
which would better utilize the hardware resources.