split: connection: fix connecting process stuck
In the `try_connect_split_device`, there is a step to check if a certain address
is already have a connection with the dongle. If yes, we will skip processing
with the peripheral. This is a speed up process for, however, zephyr BLE stack
could report that such connection exists, yet we may just disconnect with that
peripheral. As such, the whole connecting process is terminated without
establishing the connection, thus the following connected, security changed
calls are not executed, thus no new start scan is triggered.

This could happen when a peripheral is quickly turned on and off. In reality, we
can just create a new connection, and restart the whole discovery process. As a
result, the speed up process is removed.